@@ -16,7 +16,7 @@
 
 <script>
 export default {
-  name: 'Intro'
+  name: 'IntroComponent'
 }
 </script>
 
